energy_predictor/requirements.txt
Flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
plotly==5.17.0
flask-login==0.6.3
bcrypt==4.0.1
schedule==1.2.0
pymongo[srv]==4.5.0
dnspython==2.4.2